refactor to separate files and classes
