+++ frontend/index.html
@@ -40,10 +40,7 @@
     <!-- <script src="../build/index.js"></script> -->
     <!-- <script type="text/javascript" src="test.js"></script> -->
 </body>
-<!-- <script>var exports = {};</script>
-<script>var exports = {"__esModule": true};</script> -->
 <script src='https://cdn.bootcdn.net/ajax/libs/jquery/3.5.1/jquery.js'></script>
-
 <!-- <script src="../build/frontend/chesspieces.js"></script> -->
 <script src="./bundle.js"></script>
 <!-- <script src="../build/frontend/board.js"></script> -->
