--- frontend/index.ts
@@ -1,1 +1,0 @@
-// import {General} from "./chesspieces";
